hello world for server tests with jest + supertest
