+++ package.json
@@ -9,7 +9,8 @@
   },
   "scripts": {
     "start": "nodemon server/server.js",
-    "build": "webpack -d --watch"
+    "build": "webpack -d --watch",
+    "test": "jest"
   },
   "dependencies": {
     "axios": "^0.18.0",
@@ -28,7 +29,10 @@
     "babel-preset-react": "^6.24.1",
     "css-loader": "^0.28.11",
     "eslint-config-hackreactor": "git://github.com/reactorcore/eslint-config-hackreactor",
+    "jest": "^23.1.0",
     "style-loader": "^0.21.0",
+    "superagent": "^3.8.3",
+    "supertest": "^3.1.0",
     "webpack": "^4.10.2",
     "webpack-cli": "^2.1.5",
     "webpack-dev-server": "^3.1.4"
