+++ spec/client/UserInfoBox.incompletetest.js
@@ -1,0 +1,9 @@
+// const comp = require('../../client/UserInfoBox.jsx');
+
+// test.skip('adds 1 + 2 to equal 3', () => {
+//   expect(comp.snakeToCamel('hello_world_im_here')).toBe('helloWorldImHere');
+// });
+
+// test.skip('adds 1 + 2 to equal 3', () => {
+//   expect(comp.camelToSnake('helloWorldImHere')).toBe('hello_world_im_here');
+// });
